requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.0
defusedxml==0.7.1
Django==3.2.2
django-celery-beat==2.4.0
django-filter==22.1
django-social-share==2.3.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
kombu==5.2.4
oauthlib==3.2.2
Pillow==9.2.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.6.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-gettext==4.1
python-slugify==7.0.0
python3-openid==3.2.0
pytz==2022.6
redis==4.4.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
slugify==0.0.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
text-unidecode==1.3
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.5
whitenoise==2.0.6
